requirements.txt
alembic==0.9.9
Babel==2.5.3
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
click==6.7
Flask==0.12
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.4
Mako==1.0.7
MarkupSafe==0.23
passlib==1.7.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycparser==2.18
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.7
Werkzeug==0.11.15
WTForms==2.1
pyjwt
